Use single quotes around URL
This change adds single quotes around the URL of the curl command.

It makes sure to make curl commands with a URL that contains multiple query param work well in *nix terminals.

e.g. `curl http://example.com?query=one&query=two` the terminal will interpret the `&` character as a “send to background” command. [See this SO thread for details](https://stackoverflow.com/questions/13339469/how-to-include-an-character-in-a-bash-curl-statement)